requirements.txt
gunicorn
connexion[swagger-ui]
flask-cors
python-keycloak == 0.25.0
flask-sqlalchemy
psycopg2-binary
pyyaml
hvac
requests
attrs
retrying
prance
dpath
python-dateutil
cron-validator
flask-apscheduler
flask_injector
